Tucker Davis Support
At some point, a single entire channel is loaded in memory.
This is a lazy way to get the acquisition and spiking events.

Also, the importer gets TimeBounds and the sampling rates have the most ridiculous values ever, making the loading of segments to have weird lengths.
The end of the recordings is filled with zeros to satisfy the expected length.

The spiking events have not yet been tested for online clustering. Waiting for Maryse's data.