Update process_generate_fem and add possibility to generate four layer FEM mesh
Update the way to define the regions (adapted for multilayer meshing).
Add an option to ask the user to generate either 3 layers model (as it is in the previous version)
or generate 4 layers. In the case of the 4 layers, a new surface will be generated by inflating the
outer skull (function bst_inflateORdeflate_surface at line 194) by skullThiskness value (at line 336).
Update the FemMat structure in the cause of 4 layers.