Waveclus now follows the latest updates from GitHub
We used to use a fixed version of waveclus ever since we first introduced the spikesorter.
Now the updating will be done straight from the latest version of GitHub.

This might be risky in case something breaks compatibility from the Waveclus side.
However, traditionally the inputs-outputs are expected to remain the same and probably Brainstorm shouldn't be affected as is the case as of this push.