3DViz bugfix and MRI visualization for FOOOF and SPRiNT
Fixes a bug where swapping FOOOF/SPRiNT data being displayed in 3DViz would not update.

Note: Changes to mri_draw_cuts are made in the event that users try to load data from 0Hz (from FOOOF), which is a matrix of NaNs. If there are any other situations in which a matrix of NaNs could be involved in the code (which should instead cause an error), then an alternative fix will be needed.