Disable surface smoothing for Isosurface
As it is now, smoothing of the surface updates the surface using its property values stored in the database. On changing threshold of the isosurface, we are updating the surface properties in the figure but not in the database (until the user saves the changes on closing the figure) in which case smoothing won't work due to mismatch in property values for the surface in database and in the figure.